Move TestReporter to test_reporter.go
